rename: スポット詳細のコメント関連コンポーネントを整理・改名
