+++ frontend/src/components/Spot/SpotShowDialogCommentPanel.vue
@@ -1,7 +1,7 @@
 <template>
   <v-card flat outlined>
     <span v-if="spot.comments[0]">
-      <spot-show-dialog-comment-panel-index :spot="spot" :type="type" />
+      <comment-index :spot="spot" :type="type" />
     </span>
     <span v-else>
       <v-card flat class="ma-4">
@@ -11,15 +11,15 @@
 
     <v-card-actions>
       <v-spacer />
-      <spot-show-dialog-comment-panel-post-dialog :spot="spot" :type="type" />
+      <comment-post-dialog :spot="spot" :type="type" />
       <v-spacer />
     </v-card-actions>
   </v-card>
 </template>
 
 <script>
-import SpotShowDialogCommentPanelIndex from '@/components/Spot/SpotShowDialogCommentPanelIndex.vue'
-import SpotShowDialogCommentPanelPostDialog from '@/components/Spot/SpotShowDialogCommentPanelPostDialog.vue'
+import CommentIndex from '@/components/Comment/CommentIndex.vue'
+import CommentPostDialog from '@/components/Comment/CommentPostDialog.vue'
 
 export default {
   props: {
@@ -28,8 +28,8 @@ export default {
   },
 
   components: {
-    SpotShowDialogCommentPanelIndex,
-    SpotShowDialogCommentPanelPostDialog
+    CommentIndex,
+    CommentPostDialog
   }
 }
 </script>
